+++ libs/core/src/components/datepicker/datepicker.component.ts
@@ -393,7 +393,7 @@ class Datepicker extends GdsFormControlElement<Date> {
           >
             <gds-button
               @click=${this.#handleDecrementFocusedMonth}
-              aria-label=${msg('Previous month')}
+              aria-label=${msg('Switch to previous month')}
               rank="tertiary"
               size="small"
             >
@@ -438,7 +438,7 @@ class Datepicker extends GdsFormControlElement<Date> {
             </gds-dropdown>
             <gds-button
               @click=${this.#handleIncrementFocusedMonth}
-              aria-label=${msg('Next month')}
+              aria-label=${msg('Switch to next month')}
               rank="tertiary"
               size="small"
             >
@@ -482,6 +482,7 @@ class Datepicker extends GdsFormControlElement<Date> {
                         this.#dispatchInputEvent()
                         this.#dispatchChangeEvent()
                       }}
+                      aria-label=${msg('Clear selected date')}
                     >
                       ${msg('Clear')}
                     </gds-button>`,
@@ -498,6 +499,7 @@ class Datepicker extends GdsFormControlElement<Date> {
                         e.stopPropagation()
                         this.#focusDate(new Date())
                       }}
+                      aria-label=${msg("Select today's date")}
                     >
                       ${msg('Today')}
                     </gds-button>`,
@@ -873,7 +875,6 @@ class Datepicker extends GdsFormControlElement<Date> {
  * @element gds-datepicker
  * A form control that allows the user to select a date.
  *
- * @status beta
  *
  * @slot extended-supporting-text - A longer supporting text can be placed here. It will be displayed in a panel when the user clicks the info button.
  * @slot message - ***(deprecated - use `errorMessage` property instead)*** Error message to show below the input field whem there is a validation error.