+++ libs/core/src/components/dropdown/dropdown.test.ts
@@ -471,6 +471,24 @@ describe('<gds-dropdown> interactions', () => {
     await expect(changeHandler.firstCall.args[0].detail.value).to.equal('v2')
   })
 
+  it('should clear when clear button is clicked', async () => {
+    const el = await fixture<GdsDropdown>(html`
+      <gds-dropdown clearable value="v2">
+        <gds-option isplaceholder>None</gds-option>
+        <gds-option value="v1">Option 1</gds-option>
+        <gds-option id="option2" value="v2">Option 2</gds-option>
+        <gds-option value="v3">Option 3</gds-option>
+      </gds-dropdown>
+    `)
+    await aTimeout(0)
+
+    const clearButton =
+      el.shadowRoot!.querySelector<HTMLElement>('[id="clear-btn"]')!
+    await clickOnElement(clearButton)
+
+    await waitUntil(() => el.value === undefined)
+  })
+
   // Disabled because this test is unreliable. "sendMouse" does not produce a correct PointerEvent event.
   // it('should close on click outside', async () => {
   //   const el = await fixture<GdsDropdown>(html`