+++ libs/core/src/components/icon/icons/chevron-top-small.ts
@@ -1,0 +1,5 @@
+import { IconChevronTopSmall } from './chevron-top-small.component'
+
+IconChevronTopSmall.define()
+
+export { IconChevronTopSmall }