+++ libs/core/src/components/icon/icons/jpg.ts
@@ -1,0 +1,5 @@
+import { IconJpg } from './jpg.component'
+
+IconJpg.define()
+
+export { IconJpg }