+++ libs/react/src/lib/icon/icons/IconShield.tsx
@@ -1,0 +1,1 @@
+export { IconShield } from '@sebgroup/green-core/react';