+++ libs/react/src/lib/icon/icons/IconTargetArrow.tsx
@@ -1,0 +1,1 @@
+export { IconTargetArrow } from '@sebgroup/green-core/react';